refactor(flags, command): move errors to separate file
